670 - Move, rename
