+++ server/libs/modules/components/google/google-drive/src/main/java/com/bytechef/component/google/drive/GoogleDriveComponentHandler.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package com.bytechef.component.googledrive;
+package com.bytechef.component.google.drive;
 
-import static com.bytechef.component.googledrive.connection.GoogleDriveConnection.CONNECTION_DEFINITION;
-import static com.bytechef.component.googledrive.constant.GoogleDriveConstants.GOOGLEDRIVE;
+import static com.bytechef.component.google.drive.connection.GoogleDriveConnection.CONNECTION_DEFINITION;
+import static com.bytechef.component.google.drive.constant.GoogleDriveConstants.GOOGLE_DRIVE;
 import static com.bytechef.hermes.component.definition.ComponentDSL.component;
 
-import com.bytechef.component.googledrive.action.GoogleDriveUploadAction;
+import com.bytechef.component.google.drive.action.GoogleDriveUploadAction;
 import com.bytechef.hermes.component.ComponentHandler;
 import com.bytechef.hermes.component.definition.ComponentDefinition;
 import com.google.auto.service.AutoService;
@@ -31,10 +31,10 @@ import com.google.auto.service.AutoService;
 @AutoService(ComponentHandler.class)
 public class GoogleDriveComponentHandler implements ComponentHandler {
 
-    private static final ComponentDefinition COMPONENT_DEFINITION = component(GOOGLEDRIVE)
+    private static final ComponentDefinition COMPONENT_DEFINITION = component(GOOGLE_DRIVE)
         .title("Google Drive")
         .description("Component description.")
-        .icon("path:assets/googledrive.svg")
+        .icon("path:assets/google-drive.svg")
         .connection(CONNECTION_DEFINITION)
         .actions(GoogleDriveUploadAction.ACTION_DEFINITION);
