+++ server/libs/modules/components/google/google-drive/src/main/resources/files/examplePhoto.jpg
