+++ settings.gradle.kts
@@ -148,7 +148,7 @@ include("server:libs:modules:components:email")
 include("server:libs:modules:components:example")
 include("server:libs:modules:components:file-storage")
 include("server:libs:modules:components:filesystem")
-include("server:libs:modules:components:googledrive")
+include("server:libs:modules:components:google:google-drive")
 include("server:libs:modules:components:http-client")
 include("server:libs:modules:components:hubspot")
 include("server:libs:modules:components:html-helper")
